requirements.txt
accelerate==1.6.0
av==14.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
chess==1.11.2
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
peft==0.15.1
pillow==11.1.0
psutil==7.0.0
pyparsing==3.2.3
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.2
typing_extensions==4.13.2
urllib3==2.4.0
